Make mock api server pick up values from env
Makes the mock api server pick up values for HOST and PORT from the
env var PQ_REST_API_HOST, such that when it runs via docker the host
can be something other than "localhost". With this way it would be
easier to use docker linking hostnames, or arbitrary hosts.