Adds explicit logging and removes warning
Sometimes we get this warning:

```
. (called from block in tsort_each at /usr/local/lib/ruby/2.3.0/tsort.rb:228)
DEPRECATION WARNING: You did not specify a `log_level` in `production.rb`. Currently, the default value for `log_level` is `:info` for the production environment and `:debug` in all other environments. In Rails 5 the default value will be unified to `:debug` across all environments. To preserve the current setting, add the following line to your `production.rb`:

   config.log_level = :info
```
so now we specify an explicit log level for each environment.