Amend rake task for prefixing old questions
The API from Parliament does not treat each questions UIN as unique over time, instead the numbers reset at the beginning of a new Parliament. To avoid clash, though not elegantly, we prefix with a different character for each new Parliament. The application is due to be superceeded so this approach is effective.