Fix url in email and sending host regex
The URL generated for the emails included a port number from the
ActionBaseMailer settings in the config environments, this removes
where it is set.

Also fixes an issue where HostEnv.is_live? returns true for development
due to the regex ignoring the development prefix in the url. This caused
the migrations to error when deploying to development.