CT-3650 Replace ActiveRecord::Base.connected?
This method no longer reliably returns "true" when
asking if the db is visible. If there has been no
query within a certain timeframe, the connection is
released and this method returns false although
any query will reconnect

We can use the newer
ActiveRecord::Base.connection.active? which doesn't
have this behaviour and returns true if the db
can be accessed, not "is connected"