Adds new expect statement to test to enable it to pass.
Looks like updating addressable dependency has changed the way that the
mock object NotifyPqMailer was recieving calls. After the update it was
expecting that all email calls needed to be checked for rather than just
allowing it to pass even if there were calls that you didn't want to
check for.

After the change it expected you to be explicit, which I have now
changed it to check for.

looking at EarlyBirdReportService you can see that the pqtest@... email
is added to the list of recipents - the test was failing, saying that
this was now expected. Looks like it was always there but not explicitly
expected in the test.

I've now added an expectation for this and the tests pass.