Bump Rails and dependencies
Bump Rails to the latest patch in the 6.1.x branch, along with several other dependencies to their minor bugfix versions.

A few other chores I noticed as I was setting up the app locally:

* Tweaked the README and `env.sample` so it is more clear how to get the app running locally.

* Removed unused `.travis.yml`

* Removed `selenium-webdriver` because it was not being used, and Nokogiri explicit declaration because it is already part of other gems so no need for having it in the Gemfile.

* Made default `rake` command run rubocop in addition to the specs, otherwise is just too easy to forget about rubocop.

* Updated `dependabot.yml` to only raise PR for security versions, and not for version updates, as per documentation here:

https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#open-pull-requests-limit

NOTE: I tried to bump Rubocop even to the next minor release (0.89) and it went crazy with offenses, so I believe it is better to bump rubocop and fix all those offenses in a separate PR so it is more self-contained.